requirements.txt
fastapi
uvicorn[standard]
openai
requests
faiss-cpu
sentence-transformers
torch
transformers
